Lots of refactor and spelling fixes. Add quit to desktop mode.
